Move SearchParameter interface
