+++ src/main/java/com/github/thesench/solr/dsl/stream/expr/params/RawSearchParameter.java
@@ -1,7 +1,5 @@
 package com.github.thesench.solr.dsl.stream.expr.params;
 
-import com.github.thesench.solr.dsl.stream.expr.sources.SearchParameter;
-
 import org.apache.solr.client.solrj.io.stream.expr.StreamExpressionNamedParameter;
 
 public class RawSearchParameter extends StreamExpressionNamedParameter implements SearchParameter {
