+++ src/main/java/com/github/thesench/solr/dsl/stream/expr/params/Rows.java
@@ -1,7 +1,5 @@
 package com.github.thesench.solr.dsl.stream.expr.params;
 
-import com.github.thesench.solr.dsl.stream.expr.sources.SearchParameter;
-
 import org.apache.solr.client.solrj.io.stream.expr.StreamExpressionNamedParameter;
 import org.apache.solr.common.params.CommonParams;
