+++ src/main/java/com/github/thesench/solr/dsl/stream/expr/params/SearchParameter.java
@@ -1,4 +1,4 @@
-package com.github.thesench.solr.dsl.stream.expr.sources;
+package com.github.thesench.solr.dsl.stream.expr.params;
 
 import org.apache.solr.client.solrj.io.stream.expr.StreamExpressionParameter;
